requirements.txt
accelerate==1.5.2
acme==3.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.8.0
appdirs==1.4.4
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
attrs==25.1.0
audioop-lts==0.2.1
audioread==3.0.1
Authlib==1.5.1
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
build==1.2.2.post1
buildozer==1.5.0
cachelib==0.9.0
cachetools==5.5.1
cairocffi==1.7.1
CairoSVG==2.0.3
certbot==3.3.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7
configobj==5.0.9
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
db==0.1.1
deap==1.4.2
decorator==4.4.2
DeepXDE==1.13.2
Deprecated==1.2.18
diffusers==0.32.2
distlib==0.3.9
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
dnspython==2.7.0
docutils==0.21.2
elevenlabs==1.54.0
email-validator==2.1.0
et_xmlfile==2.0.0
exception==0.1.0
fake-useragent==2.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
Flask-APScheduler==1.13.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Compress==1.17
Flask-Cors==5.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-SSLify==0.1.5
flask-talisman==1.1.0
Flask-WTF==1.2.2
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-speech==2.30.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
gps==3.19
gpsd-py3==0.3.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.28.1
idna==2.9
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
instagram-private-api==1.6.0.0
instaloader==4.14.1
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
josepy==1.15.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
lightgbm==4.5.0
limits==4.0.1
llvmlite==0.44.0
lxml==5.3.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mllib==1.0.0a2
mltk==0.0.5
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.2.4
openai==1.68.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
ordered-set==4.1.0
osmnx==2.0.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
prettytable==0.7.2
proglog==0.1.10
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pyaes==1.6.1
pyaml==25.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyogrio==0.10.0
pyOpenSSL==25.0.0
pyotp==2.9.0
pyparsing==3.2.1
pypiwin32==223
pyproj==3.7.0
pyproject_hooks==1.2.0
pyreadline==2.1
pyreadline3==3.5.4
pyRFC3339==2.0.1
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-for-android==2024.1.21
python-socketio==5.12.1
python-telegram-bot==20.3
pytils==0.4.1
pytz==2025.1
pywin32==308
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-futures==1.0.2
requests-toolbelt==0.9.1
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.15.1
selenium==4.30.0
setuptools==78.1.0
sh==2.2.2
shapely==2.0.7
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spark==0.2.1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
statsmodels==0.14.4
stem==1.8.2
stopit==1.1.2
sympy==1.13.1
telegram==0.0.1
Telethon==1.39.0
threadpoolctl==3.5.0
tinycss==0.4
tokenizers==0.21.0
toml==0.10.2
tools==0.1.9
torch==2.6.0
TPOT==0.12.2
tqdm==4.67.1
transformers==4.48.3
trio==0.29.0
trio-websocket==0.12.2
twilio==9.4.6
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.3.0
v==1
virtualenv==20.29.3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xgboost==2.1.4
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
